requirements.txt
alabaster==1.0.0
anyio==4.7.0
asgiref==3.8.1
babel==2.16.0
blacknoise==1.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.4
django-rest-framework==0.1.0
djangorestframework==3.15.2
docutils==0.21.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.6
packaging==24.2
pluggy==1.5.0
Pygments==2.18.0
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.2
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvicorn-worker==0.2.0
